requirements.txt
annotated-types==0.7.0
anyio==4.3.0
certifi==2024.2.2
distro==1.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
humanize==4.9.0
idna==3.7
markdown-it-py==3.0.0
mdurl==0.1.2
openai==1.30.1
psutil==5.9.8
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
python-dotenv==1.0.1
rich==13.7.1
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.11.0
